noteworth/requirements.txt
  
astroid==2.0.4
certifi==2018.8.24
chardet==3.0.4
Django==2.2.10
django-cors-headers==2.4.0
djangorestframework==3.9.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
pylint==2.1.1
python3-openid==3.1.0
six==1.11.0
sqlparse==0.3.0
